backend/requirements.txt
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-address==0.2.1
django-allauth==0.41.0
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-mptt-admin==1.0.1
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-smartfields==1.1.1
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
phonenumbers==8.11.5
Pillow==7.0.0
python3-openid==3.1.0
pytils==0.3
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
